fix: move exported utils out of component files to satisfy react-refresh/only-export-components
