+++ src/components/Panels/itemUtils.ts
@@ -1,0 +1,3 @@
+export function formatPanelHour(hour?: Date): string {
+    return `${hour?.getHours()}:${hour?.getMinutes().toString().padStart(2, "0")}`;
+}
